Working on adding decoding animations for some reasons
It's broken and weird, maybe I need a new technique, it is harder than I thought it would be, why am I doing this?... Am I really scared of AI or something? It's broken rn and I'm tired so I'll either fix it later or abandon it.
NEXT: Maybe actually implement back propagation so this project can be over?..????
OR: DO ANOTHER STUPID TASK TO MAKE THE APP MORE DOPE BUT STILL USELESS <3 <3 <3
Goodnight.