Added glow when the player square is found
The player's square now glows when found and the other pieces are highlighted instead of flashed.
NEXT: Probabally session management stuff, maybe some map stuff but things are lookin pretty gooood!
THEN: Trainer to automatically select a map and run the AI on it to evaluate it, the chose another map and do it again, hopefully with some (Optional) user input along the way!